models_v2/pytorch/unetpp/inference/gpu/requirements-no-deps.txt
pretrainedmodels==0.7.4
efficientnet-pytorch==0.7.1
timm==0.4.12
segmentation-models-pytorch